zerodev validator & executor

+++ src/core/FeeManager.sol
@@ -7,7 +7,8 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 contract FeeManager is Ownable {
     uint256 public fee = 2000;
     bool initialized;
-    event FeePaid(address indexed from, uint amount);
+
+    event FeePaid(address indexed from, uint256 amount);
 
     function initialize(address owner) external {
         require(!initialized);
